requirements.txt
flask==3.0.0
flask-cors==4.0.0
yfinance==0.2.36
python-dateutil==2.8.2
